FSG-561 Add conditional write to modulespath
* Conditional is needed to determine if we should write to the
`modulespath` location. On Ubuntu images, the path referenced by
modulespath does not exist and causes the build to hang when setup.py
fails at this stage

+++ foundry/scripts/setup.py
@@ -217,7 +217,8 @@ def setup_modules():
 
     # for CentOS 7
     modulespath = Path('/usr/share/lmod/lmod/init/.modulespath')
-    modulespath.write_text(f"""
+    if modulepath.exists():
+       modulespath.write_text(f"""
 {dirs.modulefiles}
 """)
